refactor: make sure asset dir exists

+++ plugins/obsidian/scripts/bundle.mjs
@@ -13,7 +13,8 @@ const __dirname = path.dirname(__filename);
 const srcRoot = path.resolve(__dirname, '..');
 const distDir = path.resolve(srcRoot, 'dist');
 const projectRoot = path.resolve(srcRoot, '../..');
-const assetDestDir = path.resolve(projectRoot, 'lires_server/assets/obsidian-plugin');
+const serverAssetDir = path.resolve(projectRoot, 'lires_server/assets');
+const assetDestDir = path.resolve(serverAssetDir, 'obsidian-plugin');
 
 neededFiles.forEach(file => {
     fs.copy(`./${file}`, `${distDir}/${file}`)
@@ -21,10 +22,16 @@ neededFiles.forEach(file => {
 })
 console.info('Copied needed files to dist dir');
 
+// check if path is correct
+if (!fs.existsSync(serverAssetDir)) {
+    console.error(`dist dir ${distDir} does not exist`);
+    process.exit(1);
+}
 // if assetDestDir does exist, remove it
 if (fs.existsSync(assetDestDir)) {
     console.info(`Removing ${assetDestDir}`);
     fs.removeSync(assetDestDir);
+    fs.mkdirSync(assetDestDir);
 }
 
 // copy dist to lires_server/assets/obsidian-plugin,
@@ -33,8 +40,6 @@ fs.copy(distDir, assetDestDir, {
     filter: (src, dest) => {
         return !src.endsWith('data.json');
     }
-})
-    .then(() => {
-        console.info(`Copied dist to ${assetDestDir}`);
-    })
-    .catch(err => console.error(err));
+}).then(() => {
+    console.info(`Copied dist to ${assetDestDir}`);
+}).catch(err => console.error(err));
